chore: tests for sharding and aggregation
